global exception handler and some other changes

+++ shoppingmart/src/main/java/com/ashutosh/shoppingmart/entity/Offer.java
@@ -31,7 +31,7 @@ public class Offer {
     @Enumerated(EnumType.STRING)
     private DiscountCategory discountCategory;
 
-    private double discountValue;
+    private int discountValue;
 
     private boolean enabled;
